Verify responses have the correct parent attribute; Namespace tests, move test to separate files
An example of a parent attribute expected in a respose is a transaction
response should contain a "transaction" attribute. If the attribute is missing
a SpreedlyCore::InvalidResponse error is raised. This error subclasses
SpreedlyCore::Error. An InvalidResponse can occur if there is an error
on spreedly's side or if modification is made to a payment method that
does not exist.

Tests are now in the SpreedlyCore namespace which means you don't need to
specify the SpreedlyCore:: prefix in the tests. Tests are also split up
based on their library function. PaymentMethodTest covers transactions in
a way though since you can create transactions (retain, redact, purchase, ect)
via the payment method.